Moved cast logic to hoplalib.cast
