+++ src/hopla/hoplalib/cast/__init__.py
@@ -1,0 +1,1 @@
+#!/usr/bin/env python3
